requirements-fargate.txt
# Fargate requirements - Latest versions for containerized deployment
--prefer-binary
--extra-index-url https://pypi.org/simple

# Core AWS dependencies
boto3>=1.40.0
botocore>=1.40.0

# Pydantic and configuration
pydantic>=2.11.0
pydantic-settings>=2.10.0

# AI/ML dependencies - Latest version for Fargate
strands-agents>=1.6.0

# HTTP and networking
requests>=2.31.0
httpx>=0.28.0
httpcore>=1.0.0

# Environment and utilities
python-dotenv>=1.1.0
pytz>=2023.3

# Web framework for Fargate
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
uvloop>=0.19.0
httptools>=0.6.0

# Data processing
numpy>=1.24.0

# Calendar and scheduling (for medication reminders)
ics>=0.7.0

# Development and testing (for container health checks)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
Faker>=20.0.0
freezegun>=1.2.0
moto>=4.2.0
ruff>=0.1.0
coverage>=7.3.0

# AWS CDK (for deployment)
aws-cdk-lib>=2.100.0
constructs>=10.3.0

# Utilities
werkzeug>=3.0.0
